requirements.txt
opencv-python==3.4.1.15
Pillow==5.1.0
numpy==1.15.2